requirements.txt
mysql-connector-python==8.0.28
schematics==2.1.1
configparser==5.2.0